requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.3
django-admin==2.0.2
django-ckeditor==6.7.1
django-environ==0.11.2
django-excel-response2==3.0.6
django-js-asset==2.2.0
django-six==1.0.5
docopt==0.6.2
excel-base==1.0.4
executing==2.0.1
fastjsonschema==2.19.1
gunicorn==21.2.0
idna==3.6
ipython==8.12.3
isoweek==1.3.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
screen==1.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
TimeConvert==3.0.13
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
xlwt==1.3.0
yarg==0.1.9
